+++ Backend_Vincent/src/main/java/com/venturenix/cmc/models/GroupScore.java
@@ -9,20 +9,20 @@ import java.util.Date;
 public class GroupScore {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Integer id;
-  private Integer event_id;
-  private Integer group_id;
-  private Integer question_id;
-  private Integer testcase_total;
-  private Integer testcase_pass_total;
+  private Long id;
+  private Long event_id;
+  private Long group_id;
+  private Long question_id;
+  private Long testcase_total;
+  private Long testcase_pass_total;
   private Double testcase_score_total;
   private String status;
   private LocalDateTime createddate;
-  private Integer createdby;
+  private Long createdby;
   private LocalDateTime updateddate;
-  private Integer updatedby;
+  private Long updatedby;
 
-  public GroupScore(Integer event_id, Integer group_id, Integer question_id, Integer testcase_pass_total, Double testcase_score_total, Integer testcase_total, String status, LocalDateTime createddate, Integer createdby, LocalDateTime updateddate, Integer updatedby) {
+  public GroupScore(Long event_id, Long group_id, Long question_id, Long testcase_pass_total, Double testcase_score_total, Long testcase_total, String status, LocalDateTime createddate, Long createdby, LocalDateTime updateddate, Long updatedby) {
     this.event_id = event_id;
     this.group_id = group_id;
     this.question_id = question_id;
@@ -37,51 +37,51 @@ public class GroupScore {
     
   }
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
 
-  public Integer getEventid() {
+  public Long getEventid() {
     return event_id;
   }
 
-  public void setEventid(Integer event_id) {
+  public void setEventid(Long event_id) {
     this.event_id = event_id;
   }  
 
-  public Integer getGroupid() {
+  public Long getGroupid() {
     return group_id;
   }
 
-  public void setGroupid(Integer group_id) {
+  public void setGroupid(Long group_id) {
     this.group_id = group_id;
   }
 
-  public Integer getQuestionid() {
+  public Long getQuestionid() {
     return question_id;
   }
 
-  public void setQuestionid(Integer question_id) {
+  public void setQuestionid(Long question_id) {
     this.question_id = question_id;
   }  
 
-  public Integer getTestcasetotal() {
+  public Long getTestcasetotal() {
     return testcase_total;
   }
 
-  public void setTestcasetotal(Integer testcase_total) {
+  public void setTestcasetotal(Long testcase_total) {
     this.testcase_total = testcase_total;
   }  
 
-  public Integer getTestcasepasstotal() {
+  public Long getTestcasepasstotal() {
     return testcase_pass_total;
   }
 
-  public void setTestcasepasstotal(Integer testcase_pass_total) {
+  public void setTestcasepasstotal(Long testcase_pass_total) {
     this.testcase_pass_total = testcase_pass_total;
   } 
 
@@ -102,11 +102,11 @@ public class GroupScore {
     this.status = status;
   }
 
-  public Integer getCreatedby() {
+  public Long getCreatedby() {
     return createdby;
   }
 
-  public void setCreatedby(Integer createdby) {
+  public void setCreatedby(Long createdby) {
     this.createdby = createdby;
   }  
 
@@ -118,11 +118,11 @@ public class GroupScore {
     this.createddate = createddate;
   }  
 
-  public Integer getUpdatedby() {
+  public Long getUpdatedby() {
     return updatedby;
   }
 
-  public void setUpdatedby(Integer updatedby) {
+  public void setUpdatedby(Long updatedby) {
     this.updatedby = updatedby;
   }
