+++ Backend_Vincent/src/main/java/com/venturenix/cmc/models/TestCase.java
@@ -9,18 +9,18 @@ import java.util.Date;
 public class TestCase {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Integer id;
-  private Integer questionid;
-  private Integer testcasescoreid;
+  private Long id;
+  private Long questionid;
+  private Long testcasescoreid;
   private String testcasetext;
   private String testcaseresult;
   private String status;
   private LocalDateTime createddate;
-  private Integer createdby;
+  private Long createdby;
   private LocalDateTime updateddate;
-  private Integer updatedby;  
+  private Long updatedby;  
 
-  public TestCase(Integer questionid, Integer testcasescoreid, String testcasetext, String testcaseresult, String status, LocalDateTime createddate, Integer createdby, LocalDateTime updateddate, Integer updatedby) {
+  public TestCase(Long questionid, Long testcasescoreid, String testcasetext, String testcaseresult, String status, LocalDateTime createddate, Long createdby, LocalDateTime updateddate, Long updatedby) {
     this.questionid = questionid;
     this.testcasescoreid = testcasescoreid;
     this.testcasetext = testcasetext;
@@ -37,19 +37,19 @@ public class TestCase {
     this.testcasetext = testcasetext;
   }
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
 
-  public Integer getQuestionid() {
+  public Long getQuestionid() {
     return questionid;
   }
 
-  public void setQuestionid(Integer questionid) {
+  public void setQuestionid(Long questionid) {
     this.questionid = questionid;
   }  
 
@@ -69,11 +69,11 @@ public class TestCase {
     this.testcaseresult = testcaseresult;
   }
 
-  public Integer getTestcasescoreid() {
+  public Long getTestcasescoreid() {
     return testcasescoreid;
   }
 
-  public void setTestcasescoreid(Integer testcasescoreid) {
+  public void setTestcasescoreid(Long testcasescoreid) {
     this.testcasescoreid = testcasescoreid;
   }
 
@@ -86,11 +86,11 @@ public class TestCase {
     this.status = status;
   }
 
-  public Integer getCreatedby() {
+  public Long getCreatedby() {
     return createdby;
   }
 
-  public void setCreatedby(Integer createdby) {
+  public void setCreatedby(Long createdby) {
     this.createdby = createdby;
   }  
 
@@ -102,11 +102,11 @@ public class TestCase {
     this.createddate = createddate;
   }  
 
-  public Integer getUpdatedby() {
+  public Long getUpdatedby() {
     return updatedby;
   }
 
-  public void setUpdatedby(Integer updatedby) {
+  public void setUpdatedby(Long updatedby) {
     this.updatedby = updatedby;
   }
