+++ Backend_Vincent/src/main/java/com/venturenix/cmc/repository/EventGroupRepository.java
@@ -12,7 +12,7 @@ import java.util.List;
 @Repository
 public interface EventGroupRepository extends JpaRepository<EventGroup, Long> {
 
-    Optional<EventGroup> findById(int id);
+    Optional<EventGroup> findById(long id);
 
     List<EventGroup> findAll();
